requirements.txt
psycopg2-binary==2.8.5 
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.8
SQLAlchemy==1.3.18
